third_party/WebKit/LayoutTests/fast/dom/shadow/tabstop-property.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<p>This is basic behavior test for tabStop attribute.</p>
<pre id="console"></pre>
<script>
var div;

function resetAttributes(el)
{
    el.removeAttribute("tabindex");
    el.removeAttribute("tabstop");
}

function tabStopShouldBe(el, expected)
{
    shouldBe(el + ".tabStop", expected);
    shouldBeEqualToString(el + ".getAttribute('tabstop')", expected);
}

function test() {
    debug("Testing tabStop property and attribute");

    debug("Test tabStop normal assignment behavior");
    div = document.createElement("div");

    resetAttributes(div);
    var flag = div.tabStop;
    shouldBeFalse("div.tabStop");

    div.tabStop = flag;
    tabStopShouldBe("div", "false");

    div.removeAttribute('tabstop');
    shouldBeFalse("div.tabStop");
    shouldBeFalse("div.hasAttribute('tabstop')");

    div.tabStop = true;
    tabStopShouldBe("div", "true");

    div.removeAttribute('tabstop');
    shouldBeFalse("div.tabStop");
    shouldBeFalse("div.hasAttribute('tabstop')");


    debug("Test tabStop value implicitly set by tabindex property");
    resetAttributes(div);
    div.tabIndex = 0;
    shouldBeTrue("div.tabStop");

    div.tabIndex = -1;
    shouldBeFalse("div.tabStop");

    div.tabIndex = 1;
    shouldBeTrue("div.tabStop");

    resetAttributes(div);
    shouldBeFalse("div.tabStop");


    debug("Test explicit tabStop not overridable by tabindex attribute");
    resetAttributes(div);
    div.tabStop = false;
    div.setAttribute("tabindex", "0");
    tabStopShouldBe("div", "false");

    div.tabStop = true;
    div.setAttribute("tabindex", "-1");
    tabStopShouldBe("div", "true");

    div.tabStop = false;
    div.setAttribute("tabIndex", "1");
    tabStopShouldBe("div", "false");

    resetAttributes(div);
    shouldBeFalse("div.tabStop");
    shouldBeFalse("div.hasAttribute('tabstop')");


    debug("Test tabStop change with explicit tabindex change");
    resetAttributes(div);

    div.tabIndex = 0;
    shouldBeTrue("div.tabStop");
    div.tabStop = false;
    tabStopShouldBe("div", "false");
    resetAttributes(div);

    div.tabIndex = -1;
    shouldBeFalse("div.tabStop");
    div.tabStop = true;
    tabStopShouldBe("div", "true");
    resetAttributes(div);

    debug("Test tabStop with invalid value");
    resetAttributes(div);

    div.tabIndex = 0;
    div.tabStop = false;
    tabStopShouldBe("div", "false");
    div.setAttribute("tabstop", "invalid");
    shouldBeTrue("div.tabStop");
    shouldBeEqualToString("div.getAttribute('tabstop')", "invalid");

    div.tabIndex = -1;
    div.tabStop = true;
    tabStopShouldBe("div", "true");
    div.setAttribute("tabstop", "invalid");
    shouldBeFalse("div.tabStop");
    shouldBeEqualToString("div.getAttribute('tabstop')", "invalid");
}

test();
debug('Test finished.');
</script>
</body>
</html>
